Backend/requirements.txt
Django>=4.0
djangorestframework
pillow
django-cors-headers
python-decouple
